Add files via upload
This project was setup to allow me to learn the functionality and processes of Django due to lack of experience with the framework. I created custom models to add data to my remote SQL database (Postrgres) then utilised python to manipulate and display the data to the user in various forms. The site includes user authentication to add data to the database and an admin login, for site and data management.